feat(IoP): cleanup/refactor
