+++ src/Utilities/routing/compliancePaths.hcc.js
@@ -1,29 +1,13 @@
-import {
-  COMPLIANCE_POLICIES_CREATE_PATH,
-  mapCompliancePathForIop,
-} from './compliancePathMapping';
+import { stripLeadingSlash } from './compliancePathMapping';
 
-export { COMPLIANCE_POLICIES_CREATE_PATH, mapCompliancePathForIop };
-
-export function compliancePoliciesCreatePath() {
-  return COMPLIANCE_POLICIES_CREATE_PATH;
-}
-
-export function resolveComplianceRouterTo(to) {
-  return to;
-}
+const HCC_COMPLIANCE_PREFIX = 'insights/compliance/';
 
 export function normalizePathForRouteMatch(pathname = '') {
-  const stripLeadingSlash = (s) => (s && s.startsWith('/') ? s.slice(1) : s);
   let p = stripLeadingSlash((pathname || '').split('?')[0]);
 
-  const prefix = 'insights/compliance/';
-  if (p.startsWith(prefix)) {
-    p = p.slice(prefix.length);
+  if (p.startsWith(HCC_COMPLIANCE_PREFIX)) {
+    p = p.slice(HCC_COMPLIANCE_PREFIX.length);
   }
-  return p;
-}
 
-export function complianceDocumentHref(hccHref) {
-  return hccHref;
+  return p;
 }
