+++ src/Utilities/routing/compliancePaths.test.js
@@ -1,15 +1,6 @@
-import {
-  COMPLIANCE_POLICIES_CREATE_PATH,
-  mapCompliancePathForIop,
-} from './compliancePathMapping';
-import {
-  compliancePoliciesCreatePath as compliancePoliciesCreatePathHcc,
-  resolveComplianceRouterTo as resolveComplianceRouterToHcc,
-} from './compliancePaths.hcc';
-import {
-  compliancePoliciesCreatePath as compliancePoliciesCreatePathIop,
-  resolveComplianceRouterTo as resolveComplianceRouterToIop,
-} from './compliancePaths.iop';
+import { COMPLIANCE_POLICIES_CREATE_PATH } from '@/constants';
+import { mapCompliancePathForIop } from './compliancePathMapping';
+import { resolveComplianceRouterTo } from './compliancePaths.iop';
 
 describe('mapCompliancePathForIop', () => {
   it('maps reports', () => {
@@ -22,7 +13,7 @@ describe('mapCompliancePathForIop', () => {
   });
 
   it('maps scappolicies', () => {
-    expect(mapCompliancePathForIop('/scappolicies/new')).toBe(
+    expect(mapCompliancePathForIop(COMPLIANCE_POLICIES_CREATE_PATH)).toBe(
       '/insights_compliance/scappolicies/new',
     );
   });
@@ -38,32 +29,10 @@ describe('mapCompliancePathForIop', () => {
   });
 });
 
-describe('resolveComplianceRouterTo (HCC build)', () => {
-  it('returns input unchanged', () => {
-    expect(resolveComplianceRouterToHcc('/reports')).toBe('/reports');
-  });
-});
-
 describe('resolveComplianceRouterTo (IoP build)', () => {
   it('maps string paths', () => {
-    expect(resolveComplianceRouterToIop('/reports/1')).toBe(
+    expect(resolveComplianceRouterTo('/reports/1')).toBe(
       '/insights_compliance/reports/1',
-    );
-  });
-});
-
-describe('compliancePoliciesCreatePath (HCC build)', () => {
-  it('returns legacy path', () => {
-    expect(compliancePoliciesCreatePathHcc()).toBe(
-      COMPLIANCE_POLICIES_CREATE_PATH,
-    );
-  });
-});
-
-describe('compliancePoliciesCreatePath (IoP build)', () => {
-  it('returns IoP segment path', () => {
-    expect(compliancePoliciesCreatePathIop()).toBe(
-      '/insights_compliance/scappolicies/new',
     );
   });
 });
