+++ src/constants.js
@@ -2,6 +2,8 @@
 export const APP_ID = 'compliance';
 export const DEFAULT_TITLE = 'Compliance';
 
+export const COMPLIANCE_POLICIES_CREATE_PATH = '/scappolicies/new';
+
 export const KESSEL_API_BASE_URL = '/api/kessel/v1beta2';
 
 export const SEVERITY_LEVELS = ['high', 'medium', 'low', 'unknown'];
